requirements.txt
numpy>=1.24
pandas>=2.0
geopandas>=0.14
shapely>=2.0
rasterio>=1.3
pyproj>=3.6
tqdm>=4.66
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.8
pyyaml>=6.0
torch>=2.1
torchvision>=0.16
einops>=0.7
xgboost>=2.0
lightgbm>=4.0
